usar archivo de schema sql y cambio de tipo de dato en id

+++ src/main/java/com/mitarcon/connections/domain/GreetingEntity.java
@@ -1,7 +1,5 @@
 package com.mitarcon.connections.domain;
 
-import java.math.BigInteger;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
@@ -14,6 +12,6 @@ public class GreetingEntity {
 	
 	@Id
 	@GeneratedValue
-	private BigInteger id;
+	private int id;
 	private String text;
 }
