+++ src/main/java/com/mitarcon/connections/service/dto/GreetingDTO.java
@@ -1,11 +1,9 @@
 package com.mitarcon.connections.service.dto;
 
-import java.math.BigInteger;
-
 import lombok.Data;
 
 @Data
 public class GreetingDTO {
-	private BigInteger id;
+	private int id;
 	private String text;
 }
